+++ src/plonemeeting/restapi/tests/test_services_search.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 
-from DateTime import DateTime
+from datetime import datetime
+from plone.app.textfield.value import RichTextValue
+from plonemeeting.restapi.config import CONFIG_ID_ERROR
 from plonemeeting.restapi.config import CONFIG_ID_NOT_FOUND_ERROR
 from plonemeeting.restapi.config import IN_NAME_OF_UNAUTHORIZED
 from plonemeeting.restapi.config import TYPE_WITHOUT_CONFIG_ID_ERROR
@@ -33,17 +35,17 @@ class testServiceSearch(BaseTestCase):
             self.portal_url, self.meetingConfig.getId()
         )
         response = self.api_session.get(endpoint_url)
-        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.status_code, 200, response.content)
         # nothing found for now
         self.assertEqual(response.json()[u"items_total"], 0)
 
         # create 2 items
         self.changeUser("pmManager")
         item1 = self.create("MeetingItem")
-        self.assertEqual(item1.queryState(), "itemcreated")
+        self.assertEqual(item1.query_state(), "itemcreated")
         item2 = self.create("MeetingItem")
         self.validateItem(item2)
-        self.assertEqual(item2.queryState(), "validated")
+        self.assertEqual(item2.query_state(), "validated")
         transaction.commit()
 
         # found
@@ -81,7 +83,7 @@ class testServiceSearch(BaseTestCase):
         # items are returned sorted
         self.assertEqual(
             [elt["UID"] for elt in response.json()[u"items"]],
-            [obj.UID() for obj in meeting.getItems(ordered=True)],
+            [obj.UID() for obj in meeting.get_items(ordered=True)],
         )
         transaction.abort()
 
@@ -92,7 +94,7 @@ class testServiceSearch(BaseTestCase):
         self.changeUser("pmManager")
         cfg.setUseGroupsAsCategories(False)
         self.getMeetingFolder()
-        meeting = self.create("Meeting", date=DateTime("2020/06/08 08:00"))
+        meeting = self.create("Meeting", date=datetime(2020, 6, 8, 8, 0))
         item = self.create("MeetingItem")
         item.setMotivation("<p>Motivation</p>")
         item.setDecision(self.decisionText)
@@ -173,7 +175,7 @@ class testServiceSearch(BaseTestCase):
         )
         endpoint_url += "&config_id={0}".format(self.meetingConfig.getId())
         response = self.api_session.get(endpoint_url)
-        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.status_code, 200, response.content)
 
     def test_restapi_search_meetings_endpoint(self):
         """ """
@@ -181,17 +183,17 @@ class testServiceSearch(BaseTestCase):
             self.portal_url, self.meetingConfig.getId()
         )
         response = self.api_session.get(endpoint_url)
-        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.status_code, 200, response.content)
         # nothing found for now
         self.assertEqual(response.json()[u"items_total"], 0)
 
         # create 2 meetings
         self.changeUser("pmManager")
-        meeting = self.create("Meeting", date=DateTime("2018/11/18"))
-        self.assertEqual(meeting.queryState(), "created")
-        meeting2 = self.create("Meeting", date=DateTime("2019/11/18"))
+        meeting = self.create("Meeting", date=datetime(2018, 11, 18))
+        self.assertEqual(meeting.query_state(), "created")
+        meeting2 = self.create("Meeting", date=datetime(2019, 11, 18))
         self.closeMeeting(meeting2)
-        self.assertEqual(meeting2.queryState(), "closed")
+        self.assertEqual(meeting2.query_state(), "closed")
         transaction.commit()
 
         # found
@@ -218,11 +220,12 @@ class testServiceSearch(BaseTestCase):
 
         # create 2 meetings
         self.changeUser("pmManager")
-        meeting = self.create("Meeting", date=DateTime("2019/11/18"))
-        self.assertEqual(meeting.queryState(), "created")
-        meeting2 = self.create("Meeting", date=DateTime("2019/11/18"))
+        meeting = self.create("Meeting", date=datetime(2019, 11, 18))
+        self.assertEqual(meeting.query_state(), "created")
+        meeting2 = self.create("Meeting", date=datetime(2019, 11, 18))
+        meeting2.assembly = RichTextValue(u'Mr Present, [[Mr Absent]], Mr Present2')
         self.closeMeeting(meeting2)
-        self.assertEqual(meeting2.queryState(), "closed")
+        self.assertEqual(meeting2.query_state(), "closed")
         transaction.commit()
 
         # found
@@ -237,9 +240,11 @@ class testServiceSearch(BaseTestCase):
 
         # includes every data as well as extra formatted values
         self.assertTrue("date" in resp_json["items"][0])
-        self.assertTrue("startDate" in resp_json["items"][0])
+        self.assertTrue("start_date" in resp_json["items"][0])
         self.assertTrue("notes" in resp_json["items"][0])
-        self.assertTrue("formatted_assembly" in resp_json["items"][0])
+        self.assertEqual(
+            resp_json["items"][0]["formatted_assembly"],
+            u'<p>Mr Present, <strike>Mr Absent</strike>, Mr Present2</p>')
         transaction.abort()
 
     def test_restapi_search_meetings_accepting_items(self):
@@ -251,11 +256,11 @@ class testServiceSearch(BaseTestCase):
 
         # create 2 meetings
         self.changeUser("pmManager")
-        meeting = self.create("Meeting", date=DateTime("2020/05/10"))
-        self.assertEqual(meeting.queryState(), "created")
-        meeting2 = self.create("Meeting", date=DateTime("2020/05/17"))
+        meeting = self.create("Meeting", date=datetime(2020, 5, 10))
+        self.assertEqual(meeting.query_state(), "created")
+        meeting2 = self.create("Meeting", date=datetime(2020, 5, 17))
         self.closeMeeting(meeting2)
-        self.assertEqual(meeting2.queryState(), "closed")
+        self.assertEqual(meeting2.query_state(), "closed")
         transaction.commit()
         # only meeting is accepting items
         self.assertEqual(
@@ -322,7 +327,7 @@ class testServiceSearch(BaseTestCase):
         )
         response = self.api_session.get(endpoint_url)
         # nothing found for now
-        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.status_code, 200, response.content)
         self.assertEqual(response.json()[u"items_total"], 0)
         # create one item for developers and one for vendors as pmManager
         # and one item as pmCreator1